Breakpoint management Added a debug helper class, to management breaks.  Added words #addbp #rembp #togbp to add, remove and toggle breakpoints.

+++ SmallForth/InputProcessor.cpp
@@ -151,7 +151,7 @@ bool InputProcessor::Interpret(ExecState* pExecState) {
 				else {
 					pExecState->SetVariable("#postponeState", false);
 
-					int nDebugState = pExecState->GetIntTLSVariable(ExecState::c_debugStateIndex);
+					int nDebugState = (int)pExecState->GetIntTLSVariable(ExecState::c_debugStateIndex);
 					if (nDebugState >= 1 && nDebugState < 3) {
 						std::ostream* pStdout = pExecState->GetStdout();
 						(*pStdout) << "Compiling word: " << pWord->GetName() << std::endl;
